third_party/WebKit/LayoutTests/fast/canvas/canvas-text-ideographic-space-expected.txt

PASS spanWidth == canvasWidth is true
PASS successfullyParsed is true

TEST COMPLETE
a　b　c
